requirements/base.txt
django-shop>=0.9.3
django-cms>=3.3
django-admin-sortable2>=0.6.4
django-measurement>=2.4.0
django-mptt>=0.8.6
django-parler>=1.6.5
